pyproject.toml
[project]
name = "efficient_ai_planner"
version = "0.1.0"
description = "This serves as a design for training small LLMs for task planning and execution."
authors = [
    {name = "Niall Krebs-Taylor", email = "niallkrebstaylor@gmail.com"},
]
license = "Apache License 2.0"
readme = "README.md"
dependencies = [
    "absl-py>=2.2.2",
    "datasets>=3.5.1",
    "google-adk>=0.1.0",
    "grip>=4.6.2",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "litellm>=1.66.3",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "peft>=0.15.2",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "trl>=0.17.0",
]

[tool.poetry.dependencies]
python = "^3.9"
google-adk = ">=0.0.2"
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.88.0"}


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { version = ">=0.0.2", extras = ["eval"] }
pytest = "^8.3.5"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
